+++ src/main/java/com/activedge/usermgt/model/mapper/PermissionMapper.java
@@ -6,6 +6,7 @@ import com.activedge.usermgt.model.Permission;
 import com.activedge.usermgt.model.dto.PermissionDTO;
 import org.mapstruct.*;
 
+import java.math.BigInteger;
 import java.util.Set;
 
 /**
@@ -23,7 +24,7 @@ public interface PermissionMapper extends EntityMapper<PermissionDTO, Permission
 
     Set<PermissionDTO> toDtoSet(Set<Permission> permissions);
 
-    default Permission fromId(Long id) {
+    default Permission fromId(BigInteger id) {
         if (id == null) {
             return null;
         }
